refactor(pegout): remove factory class for pegout signed cache
