+++ src/main/java/co/rsk/federate/btcreleaseclient/cache/PegoutSignedCacheImpl.java
@@ -6,13 +6,19 @@ import java.time.Instant;
 import java.util.Map;
 import java.util.Optional;
 import java.util.concurrent.ConcurrentHashMap;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 class PegoutSignedCacheImpl implements PegoutSignedCache {
+
+  private static final Logger logger = LoggerFactory.getLogger(PegoutSignedCacheImpl.class);
 
   private final Map<Keccak256, Instant> cache = new ConcurrentHashMap<>();
   private final Duration ttl;
 
   PegoutSignedCacheImpl(Duration ttl) {
+    assertValidTtl(ttl);
+
     this.ttl = ttl;
   }
 
@@ -32,8 +38,19 @@ class PegoutSignedCacheImpl implements PegoutSignedCache {
 
   private boolean hasTimestampNotExpired(Instant timestampInCache) {
     return Optional.ofNullable(timestampInCache)
-      .map(timestamp -> Instant.now().toEpochMilli() - timestamp.toEpochMilli())
-      .map(timeCachedInMillis -> timeCachedInMillis <= ttl.toMillis())
-      .orElse(false);
+        .map(timestamp -> Instant.now().toEpochMilli() - timestamp.toEpochMilli())
+        .map(timeCachedInMillis -> timeCachedInMillis <= ttl.toMillis())
+        .orElse(false);
+  }
+
+  private static void assertValidTtl(Duration ttl) {
+    if (ttl == null || ttl.isNegative() || ttl.isZero()) {
+      Long ttlInMinutes = ttl != null ? ttl.toMinutes() : null;
+      String message = String.format(
+          "Invalid pegouts signed cache TTL value in minutes supplied: %d", ttlInMinutes);
+      logger.error("[assertValidTtl] {}", message);
+
+      throw new IllegalArgumentException(message);
+    }
   }
 }
